+++ api/alembic/versions/298f400dfa64_zonal_elevation_stats.py
@@ -1,0 +1,51 @@
+"""zonal elevation stats
+
+Revision ID: 298f400dfa64
+Revises: 0dd8306467a9
+Create Date: 2023-02-01 15:56:27.433120
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '298f400dfa64'
+down_revision = '0dd8306467a9'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic ###
+    op.create_table('advisory_elevation_stats',
+                    sa.Column('id', sa.Integer(), nullable=False),
+                    sa.Column('advisory_shape_id', sa.Integer(), nullable=False),
+                    sa.Column('threshold', sa.Integer(), nullable=False),
+                    sa.Column('run_parameters', sa.Integer(), nullable=False),
+                    sa.Column('minimum', sa.Float(), nullable=False),
+                    sa.Column('quartile_25', sa.Float(), nullable=False),
+                    sa.Column('median', sa.Float(), nullable=False),
+                    sa.Column('quartile_75', sa.Float(), nullable=False),
+                    sa.Column('maximum', sa.Float(), nullable=False),
+                    sa.ForeignKeyConstraint(['advisory_shape_id'], ['advisory_shapes.id'], ),
+                    sa.ForeignKeyConstraint(['run_parameters'], ['run_parameters.id'], ),
+                    sa.ForeignKeyConstraint(['threshold'], ['advisory_hfi_classification_threshold.id'], ),
+                    sa.PrimaryKeyConstraint('id'),
+                    comment='Elevation stats per fire zone by advisory threshold'
+                    )
+    op.create_index(op.f('ix_advisory_elevation_stats_advisory_shape_id'),
+                    'advisory_elevation_stats', ['advisory_shape_id'], unique=False)
+    op.create_index(op.f('ix_advisory_elevation_stats_id'), 'advisory_elevation_stats', ['id'], unique=False)
+    op.create_index(op.f('ix_advisory_elevation_stats_run_parameters'),
+                    'advisory_elevation_stats', ['run_parameters'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic ###
+    op.drop_index(op.f('ix_advisory_elevation_stats_run_parameters'), table_name='advisory_elevation_stats')
+    op.drop_index(op.f('ix_advisory_elevation_stats_id'), table_name='advisory_elevation_stats')
+    op.drop_index(op.f('ix_advisory_elevation_stats_advisory_shape_id'), table_name='advisory_elevation_stats')
+    op.drop_table('advisory_elevation_stats')
+    # ### end Alembic commands ###
