--- api/alembic/versions/65205f6cad16_zonal_elevation_stats.py
@@ -1,86 +1,0 @@
-"""zonal elevation stats
-
-Revision ID: 65205f6cad16
-Revises: 04bc7be60b4f
-Create Date: 2023-01-26 16:51:35.673519
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '65205f6cad16'
-down_revision = '04bc7be60b4f'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic ###
-    op.create_table('advisory_elevation_stats',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('advisory_shape_id', sa.Integer(), nullable=False),
-                    sa.Column('threshold', sa.Integer(), nullable=False),
-                    sa.Column('run_parameters', sa.Integer(), nullable=False),
-                    sa.Column('minimum', sa.Float(), nullable=False),
-                    sa.Column('quartile_25', sa.Float(), nullable=False),
-                    sa.Column('median', sa.Float(), nullable=False),
-                    sa.Column('quartile_75', sa.Float(), nullable=False),
-                    sa.Column('maximum', sa.Float(), nullable=False),
-                    sa.ForeignKeyConstraint(['advisory_shape_id'], ['advisory_shapes.id'], ),
-                    sa.ForeignKeyConstraint(['run_parameters'], ['run_parameters.id'], ),
-                    sa.ForeignKeyConstraint(['threshold'], ['advisory_hfi_classification_threshold.id'], ),
-                    sa.PrimaryKeyConstraint('id'),
-                    comment='Elevation stats per fire zone by advisory threshold'
-                    )
-    op.create_index(op.f('ix_advisory_elevation_stats_advisory_shape_id'),
-                    'advisory_elevation_stats', ['advisory_shape_id'], unique=False)
-    op.create_index(op.f('ix_advisory_elevation_stats_id'), 'advisory_elevation_stats', ['id'], unique=False)
-    op.create_index(op.f('ix_advisory_elevation_stats_run_parameters'),
-                    'advisory_elevation_stats', ['run_parameters'], unique=False)
-    op.create_table('advisory_shape_elevation',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('advisory_shape_id', sa.Integer(), nullable=False),
-                    sa.Column('minimum', sa.Float(), nullable=False),
-                    sa.Column('maximum', sa.Float(), nullable=False),
-                    sa.ForeignKeyConstraint(['advisory_shape_id'], ['advisory_shapes.id'], ),
-                    sa.PrimaryKeyConstraint('id'),
-                    comment='Elevation information/stats about each advisory shape'
-                    )
-    op.create_index(op.f('ix_advisory_shape_elevation_advisory_shape_id'),
-                    'advisory_shape_elevation', ['advisory_shape_id'], unique=False)
-    op.create_index(op.f('ix_advisory_shape_elevation_id'), 'advisory_shape_elevation', ['id'], unique=False)
-    op.create_index(op.f('ix_high_hfi_area_run_parameters'), 'high_hfi_area', ['run_parameters'], unique=False)
-    op.create_table_comment(
-        'high_hfi_area',
-        'Area under advisory/warning per fire zone. advisory_area refers to the total area in a fire zone with HFI values between 4000 - 10000 and warn_area refers to the total area in a fire zone with HFI values exceeding 10000.',
-        existing_comment='Area under advisory/warning per fire zone.',
-        schema=None
-    )
-    op.create_index(op.f('ix_run_parameters_for_date'), 'run_parameters', ['for_date'], unique=False)
-    op.create_index(op.f('ix_run_parameters_run_datetime'), 'run_parameters', ['run_datetime'], unique=False)
-    op.create_index(op.f('ix_run_parameters_run_type'), 'run_parameters', ['run_type'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic ###
-    op.drop_index(op.f('ix_run_parameters_run_type'), table_name='run_parameters')
-    op.drop_index(op.f('ix_run_parameters_run_datetime'), table_name='run_parameters')
-    op.drop_index(op.f('ix_run_parameters_for_date'), table_name='run_parameters')
-    op.create_table_comment(
-        'high_hfi_area',
-        'Area under advisory/warning per fire zone.',
-        existing_comment='Area under advisory/warning per fire zone. advisory_area refers to the total area in a fire zone with HFI values between 4000 - 10000 and warn_area refers to the total area in a fire zone with HFI values exceeding 10000.',
-        schema=None
-    )
-    op.drop_index(op.f('ix_high_hfi_area_run_parameters'), table_name='high_hfi_area')
-    op.drop_index(op.f('ix_advisory_shape_elevation_id'), table_name='advisory_shape_elevation')
-    op.drop_index(op.f('ix_advisory_shape_elevation_advisory_shape_id'), table_name='advisory_shape_elevation')
-    op.drop_table('advisory_shape_elevation')
-    op.drop_index(op.f('ix_advisory_elevation_stats_run_parameters'), table_name='advisory_elevation_stats')
-    op.drop_index(op.f('ix_advisory_elevation_stats_id'), table_name='advisory_elevation_stats')
-    op.drop_index(op.f('ix_advisory_elevation_stats_advisory_shape_id'), table_name='advisory_elevation_stats')
-    op.drop_table('advisory_elevation_stats')
-    # ### end Alembic commands ###
